pyproject.toml
[project]
name = "rxcalendar"
version = "0.1.0"
description = "2026 Calendar with weekday comments and JSON export"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pillow>=12.0.0",
    "reflex>=0.8.18",
    "reportlab>=4.4.4",
]
